Fix firestoreService addCraftsman function
Previously, this function did not successfully copy reviews from
an unregistered user that had been reviewed when a craftsman with
matching credentials registers. It seems as though firestore's
ArrayUnion function can only add one item per call.

To fix that, the function loops over the unregistered reviewed user's
reviews and appends each review to the newly created document in the
craftsmen collection overwriting the old entry in the process.